+++ platform/zilliqa/zns.go
@@ -9,7 +9,7 @@ type ZNSResponse struct {
 	Addresses map[string]string
 }
 
-func (p *Platform) Lookup(coins []uint64, name string) ([]blockatlas.Resolved, error) {
+func (p *Platform) Lookup(coins []int, name string) ([]blockatlas.Resolved, error) {
 	var result []blockatlas.Resolved
 	resp, err := p.udClient.LookupName(name)
 	if err != nil {
@@ -17,7 +17,7 @@ func (p *Platform) Lookup(coins []uint64, name string) ([]blockatlas.Resolved, e
 	}
 	for _, coin := range coins {
 		symbol := CoinType.Coins[uint(coin)].Symbol
-		result = append(result, blockatlas.Resolved{Coin: coin, Result: resp.Addresses[symbol]})
+		result = append(result, blockatlas.Resolved{Coin: uint64(coin), Result: resp.Addresses[symbol]})
 	}
 	return result, nil
 }
